change the uuid for rhevm cases
